update perf demo
