+++ demos/homepage/_includes/header.html
@@ -1,7 +1,8 @@
 <!-- Header -->
 <header class="mdl-layout__header mdl-layout--fixed-header">
   <script>
-    if (location.protocol !== "https:") {
+    if (location.protocol !== "https:" &&
+        location.hostname === 'deeplearnjs.org') {
       location.protocol = "https:";
     }
   </script>
